Fix outdated seenList response from Cloudinary CDN
Cloudinary's CDN can cache files for up to 30 days. This causes
outdated seenList files to be returned at the start of the app,
even with Cache-Control set to no-cache. Instead, an 'invalidate'
parameter is passed along to the upload in order to ensure
that the CDNs invalidate their copy. Also:
- An extra script parameter has been added to skip the initial
  cloudinary database load. Useful to trigger an upload.